index: Compute type from a cache map.
Previously we cached the type as a field __type on the object
in a shared getter defined as a static property on the class.

Now we cache explicitely in a Map. The keyspace of unique class
names for errors should be reasonably bounded and we have a little
function that can read/write from this cache map.